moved all database management functions to a separate database_api module
